Implemented PlayFriendModal
