+++ assets/js/components/FriendButtons.js
@@ -4,7 +4,7 @@ class FriendButtons extends AbstractComponent {
   mount() {
     this.el.children.item(0).addEventListener('click', async (event) => {
       event.preventDefault();
-      this.props.playFriendModal.modal.show();
+      this.props.playFriendModal.props.modal.show();
     });
 
     this.el.children.item(1).addEventListener('click', async (event) => {
